refactor(orders): split create page into form hooks and step components
Extracts the 843-line create page into useCreateOrderForm/useOrderProducts
hooks and per-step presentational components (school/client/products/order),
bringing the entrypoint and every component under the size limits. Refs #81.
