+++ resources/js/pages/orders/components/product-list-item.tsx
@@ -1,0 +1,92 @@
+import { Badge } from '@/components/ui/badge';
+import { Button } from '@/components/ui/button';
+import {
+    Edit,
+    RectangleHorizontal,
+    RectangleVertical,
+    Square,
+    Trash,
+    User,
+    Users,
+} from 'lucide-react';
+import { OrderFormData } from '../form-state';
+
+interface ProductListItemProps {
+    detail: OrderFormData['order_details'][number];
+    product: Product;
+    combo?: Combo;
+    index: number;
+    onEdit: (index: number) => void;
+    onRemove: (index: number) => void;
+}
+
+export function ProductListItem({
+    detail,
+    product,
+    combo,
+    index,
+    onEdit,
+    onRemove,
+}: ProductListItemProps) {
+    return (
+        <li className="flex items-center justify-between rounded-md border border-input bg-background px-4 py-2">
+            <div className="flex flex-col gap-2">
+                <span>
+                    {product.name}
+                    {combo ? ` (combo ${combo.name})` : ''}
+                </span>
+                {product.product_type_id === 1 ? (
+                    <div className="flex items-center">
+                        <Badge variant="outline" className="gap-1 rounded-lg">
+                            <Square
+                                className="h-4 w-4"
+                                style={{ fill: detail.variant?.color }}
+                            />
+                        </Badge>
+                        <Badge variant="outline" className="rounded-lg">
+                            {detail.variant?.background}
+                        </Badge>
+                        <Badge variant="outline" className="rounded-lg">
+                            {detail.variant?.photo_type === 'individual' ? (
+                                <User className="h-4 w-4" />
+                            ) : (
+                                <Users className="h-4 w-4" />
+                            )}
+                        </Badge>
+                        <Badge variant="outline" className="rounded-lg">
+                            {detail.variant?.orientation === 'vertical' ? (
+                                <RectangleVertical className="h-4 w-4" />
+                            ) : (
+                                <RectangleHorizontal className="h-4 w-4" />
+                            )}
+                        </Badge>
+                    </div>
+                ) : undefined}
+            </div>
+            <div className="flex gap-1">
+                <Button
+                    variant="warning"
+                    size="icon"
+                    title="Editar variantes y notas"
+                    onClick={(e) => {
+                        e.preventDefault();
+                        onEdit(index);
+                    }}
+                >
+                    <Edit className="h-4 w-4" />
+                </Button>
+                <Button
+                    variant="destructive"
+                    size="icon"
+                    title="Quitar del pedido"
+                    onClick={(e) => {
+                        e.preventDefault();
+                        onRemove(index);
+                    }}
+                >
+                    <Trash className="h-4 w-4" />
+                </Button>
+            </div>
+        </li>
+    );
+}
