+++ resources/js/pages/orders/create.tsx
@@ -1,55 +1,14 @@
-import { DatePicker } from '@/components/date-picker';
-import InputError from '@/components/input-error';
-import InputHint from '@/components/input-hint';
-import {
-    Accordion,
-    AccordionContent,
-    AccordionItem,
-    AccordionTrigger,
-} from '@/components/ui/accordion';
-import { Badge, badgeVariants } from '@/components/ui/badge';
-import { Button } from '@/components/ui/button';
-import { Combobox } from '@/components/ui/combobox';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-import {
-    Select,
-    SelectContent,
-    SelectItem,
-    SelectTrigger,
-    SelectValue,
-} from '@/components/ui/select';
+import { Accordion } from '@/components/ui/accordion';
 import AppLayout from '@/layouts/app-layout';
-import { cn, formatPrice } from '@/lib/utils';
 import { BreadcrumbItem } from '@/types';
-import { Head, Link, useForm } from '@inertiajs/react';
-import axios from 'axios';
-import { format } from 'date-fns';
-import {
-    AlertCircle,
-    Edit,
-    PlusIcon,
-    RectangleHorizontal,
-    RectangleVertical,
-    Square,
-    Trash,
-    User,
-    Users,
-} from 'lucide-react';
-import { FormEvent, FormEventHandler, useEffect, useState } from 'react';
-import { toast } from 'sonner';
+import { Head } from '@inertiajs/react';
 import { AddDetail } from './add-detail';
-import { ProductOrder } from './form';
-import {
-    DraftProp,
-    OrderFormData,
-    removeDetailAt,
-    replaceDetailAt,
-    resetForNextClient,
-    resolveInitialOrderForm,
-} from './form-state';
-type SchoolLevel = 'Todos' | 'Jardin' | 'Primaria' | 'Secundaria';
-type AccordionValue = 'schools' | 'products' | 'client' | 'order' | undefined;
+import { ClientStep } from './components/client-step';
+import { OrderStep } from './components/order-step';
+import { ProductsStep } from './components/products-step';
+import { SchoolStep } from './components/school-step';
+import { DraftProp } from './form-state';
+import { SchoolLevel, useCreateOrderForm } from './hooks/use-create-order-form';
 
 const breadcrumbs: BreadcrumbItem[] = [
     {
@@ -75,197 +34,15 @@ export default function CreateOrder({
     products: Product[];
     draft?: DraftProp | null;
 }>) {
-    const [initial] = useState(() =>
-        resolveInitialOrderForm(
-            draft,
-            products.map((product) => product.id),
-        ),
-    );
-
-    const [selectedSchool, setSelectedSchool] = useState<number>(
-        initial.selectedSchool,
-    );
-    const [levelFilter, setLevelFilter] = useState<SchoolLevel>('Todos');
-
-    const [comboDropdownOpen, setComboDropdownOpen] = useState(false);
-    const [productDropdownOpen, setProductDropdownOpen] = useState(false);
-
-    const [accordionValue, setAccordionValue] =
-        useState<AccordionValue>('schools');
-
-    const [openAddModal, setOpenAddModal] = useState<
-        (Product & { combo_id?: number })[] | null
-    >(null);
-
-    const [editingIndex, setEditingIndex] = useState<number | null>(null);
-
-    const { data, setData, post, processing, errors, clearErrors } =
-        useForm<OrderFormData>(initial.form);
-
-    useEffect(() => {
-        if (initial.message) {
-            toast.info(initial.message);
-        }
-
-        if (initial.droppedProducts > 0) {
-            toast.warning(
-                `Se quitaron ${initial.droppedProducts} producto(s) del pedido recuperado porque ya no existen. Revisá el precio final.`,
-            );
-        }
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, []);
-
-    const _selectedSchool = schools.find((s) => s.id === selectedSchool);
-    const _selectedClassroom = _selectedSchool?.classrooms.find(
-        (c) => c.id === data.classroom_id,
-    );
-
-    const filteredSchools = schools.filter(
-        (school) => levelFilter === 'Todos' || school.level === levelFilter,
-    );
-
-    const counts = data.order_details.reduce<{
-        combos: Record<number, number>;
-        undefinedCount: number;
-    }>(
-        (acc, order) => {
-            if (order.combo_id !== undefined) {
-                acc.combos[order.combo_id] =
-                    (acc.combos[order.combo_id] || 0) + 1;
-            } else {
-                acc.undefinedCount++;
-            }
-            return acc;
-        },
-        { combos: {}, undefinedCount: 0 },
-    );
-
-    const error_count: Record<Exclude<AccordionValue, undefined>, boolean> = {
-        client: !!errors.name || !!errors.phone,
-        order:
-            !!errors.total_price || !!errors.payment_plan || !!errors.due_date,
-        products: !!errors.order_details,
-        schools: !!errors.classroom_id,
-    };
-
-    const toStep = (newAccordionValue: AccordionValue) => {
-        return (e: FormEvent) => {
-            e.preventDefault();
-
-            if (newAccordionValue === accordionValue) {
-                return setAccordionValue(undefined);
-            }
-
-            setAccordionValue(newAccordionValue);
-        };
-    };
-
-    const submit = (e: FormEvent, saveAndContinue = false) => {
-        e.preventDefault();
-
-        post(
-            route('orders.store', {
-                _query: {
-                    redirectTo: route(
-                        saveAndContinue ? 'orders.create' : 'orders.index',
-                    ),
-                },
-            }),
-            {
-                onSuccess: () => {
-                    toast.success('Pedido guardado con éxito');
-                    if (saveAndContinue) {
-                        setData(resetForNextClient(data));
-                        setAccordionValue('client');
-                        toast.info(
-                            'Se conservaron la escuela, las cuotas y el vencimiento. Cargá el cliente y los productos.',
-                        );
-                    }
-                },
-            },
-        );
-    };
-
-    const handleSaveAndContinue: FormEventHandler = (e) => {
-        submit(e, true);
-    };
-
-    const handleSaveAsDraft: FormEventHandler = async (e) => {
-        e.preventDefault();
-
-        try {
-            await axios.post(route('drafts.store'), {
-                classroom_id: data.classroom_id,
-                child_name: data.child_name,
-                client_name: data.name,
-                client_phone: data.phone,
-                attended_photo_session: data.attended_photo_session,
-                total_price: data.total_price,
-                payment_plan: data.payment_plan,
-                due_date: data.due_date,
-                products: data.order_details,
-            });
-
-            toast.success('Borrador guardado exitosamente');
-            window.location.href = route('drafts.index');
-        } catch {
-            toast.error('Error al guardar el borrador');
-        }
-    };
-
-    const handleAddProduct = (id: number) => {
-        setOpenAddModal([products.find((p) => p.id === id)!]);
-    };
-
-    const handleAddCombo = (id: number) => {
-        const combo = combos.find((p) => p.id === id)!;
-
-        setData(
-            'total_price',
-            String(Number(data.total_price) + Number(combo.suggested_price)),
-        );
-        setOpenAddModal(combo.products.map((p) => ({ ...p, combo_id: id })));
-    };
-
-    const setProductsOrder = (productsOrder: ProductOrder[]) => {
-        if (editingIndex !== null) {
-            setData(
-                'order_details',
-                replaceDetailAt(
-                    data.order_details,
-                    editingIndex,
-                    productsOrder,
-                ),
-            );
-            setEditingIndex(null);
-            return;
-        }
-
-        setData('order_details', [...data.order_details, ...productsOrder]);
-    };
-
-    const handleEditProduct = (index: number) => {
-        const selected = data.order_details[index];
-        // Combo items must keep the pivot-restricted variants, not the
-        // full catalog ones
-        const combo =
-            selected.combo_id !== undefined
-                ? combos.find((c) => c.id === selected.combo_id)
-                : undefined;
-        const product =
-            combo?.products.find((p) => p.id === selected.product_id) ??
-            products.find((p) => p.id === selected.product_id);
-
-        if (!product) return;
-
-        setEditingIndex(index);
-        setOpenAddModal([{ ...product, combo_id: selected.combo_id }]);
-    };
-
-    const handleRemoveProduct = (index: number) => {
-        setData('order_details', removeDetailAt(data.order_details, index));
-        toast.info('Producto quitado. Recordá revisar el precio final.');
-    };
+    const form = useCreateOrderForm({ products, combos, schools, draft });
+    const {
+        openAddModal,
+        editingIndex,
+        data,
+        setProductsOrder,
+        setOpenAddModal,
+        setEditingIndex,
+    } = form;
 
     return (
         <AppLayout breadcrumbs={breadcrumbs}>
@@ -287,555 +64,25 @@ export default function CreateOrder({
                 />
             ) : undefined}
 
-            <form onSubmit={submit} className="p-6">
+            <form onSubmit={form.submit} className="p-6">
                 <Accordion
                     type="single"
                     collapsible
                     className="w-full"
-                    value={accordionValue}
+                    value={form.accordionValue}
                 >
-                    <AccordionItem value="schools">
-                        <AccordionTrigger onClick={toStep('schools')}>
-                            <div className="flex items-center gap-2">
-                                {error_count['schools'] && (
-                                    <AlertCircle className="h-5 w-5 stroke-destructive" />
-                                )}
-                                Escuela
-                                {_selectedSchool ? (
-                                    <Badge>
-                                        {`${_selectedSchool.name} ${_selectedClassroom ? `"${_selectedClassroom.name.toUpperCase()}"` : ''}`}
-                                    </Badge>
-                                ) : undefined}
-                            </div>
-                        </AccordionTrigger>
-                        <AccordionContent className="flex flex-col gap-3 px-1">
-                            <div className="flex gap-2 py-2">
-                                Niveles:
-                                {schoolLevels.map((level) => (
-                                    <button
-                                        key={level}
-                                        onClick={(e) => {
-                                            e.preventDefault();
-                                            setLevelFilter(level);
-                                        }}
-                                        className={cn(
-                                            badgeVariants({
-                                                variant: 'outline',
-                                            }),
-                                            levelFilter === level &&
-                                                'border border-primary',
-                                        )}
-                                    >
-                                        {level}
-                                    </button>
-                                ))}
-                            </div>
-                            <div>
-                                <Label htmlFor="schoolId">Escuela</Label>
-
-                                <Select
-                                    value={String(
-                                        schools.find(
-                                            (s) => s.id === selectedSchool,
-                                        )?.id ?? '',
-                                    )}
-                                    name="schoolId"
-                                    onValueChange={(value) =>
-                                        setSelectedSchool(Number(value))
-                                    }
-                                >
-                                    <SelectTrigger>
-                                        <SelectValue />
-                                    </SelectTrigger>
-                                    <SelectContent>
-                                        {filteredSchools.map((school) => (
-                                            <SelectItem
-                                                value={String(school.id)}
-                                                key={school.id}
-                                            >
-                                                {school.name}
-                                                <Badge className="ml-1">
-                                                    {school.level}
-                                                </Badge>
-                                            </SelectItem>
-                                        ))}
-                                    </SelectContent>
-                                </Select>
-                            </div>
-
-                            <div>
-                                <Label htmlFor="classroom_id">Curso</Label>
-
-                                <Select
-                                    value={String(data.classroom_id)}
-                                    disabled={selectedSchool === 0}
-                                    name="classroom_id"
-                                    onValueChange={(value) => {
-                                        clearErrors('classroom_id');
-                                        setData('classroom_id', Number(value));
-                                    }}
-                                >
-                                    <SelectTrigger>
-                                        <SelectValue />
-                                    </SelectTrigger>
-                                    <SelectContent>
-                                        {selectedSchool !== 0 &&
-                                            schools
-                                                .find(
-                                                    (school) =>
-                                                        school.id ===
-                                                        selectedSchool,
-                                                )!
-                                                .classrooms.map((classroom) => (
-                                                    <SelectItem
-                                                        value={String(
-                                                            classroom.id,
-                                                        )}
-                                                        key={classroom.id}
-                                                    >
-                                                        {classroom.name.toUpperCase()}
-                                                    </SelectItem>
-                                                ))}
-                                    </SelectContent>
-                                </Select>
-
-                                <InputError
-                                    message={errors.classroom_id}
-                                    className="mt-2"
-                                />
-                            </div>
-
-                            <div className="mt-6 flex flex-col justify-end gap-3 md:flex-row">
-                                <Button onClick={toStep('client')}>
-                                    Siguiente
-                                </Button>
-                            </div>
-                        </AccordionContent>
-                    </AccordionItem>
-
-                    <AccordionItem value="client">
-                        <AccordionTrigger onClick={toStep('client')}>
-                            <div className="flex items-center gap-2">
-                                {error_count['client'] && (
-                                    <AlertCircle className="h-5 w-5 stroke-destructive" />
-                                )}
-                                Cliente
-                                {data.name && <Badge>{`${data.name}`}</Badge>}
-                                {data.phone && <Badge>{`${data.phone}`}</Badge>}
-                            </div>
-                        </AccordionTrigger>
-                        <AccordionContent className="px-1">
-                            <div>
-                                <Label>Nombre</Label>
-                                <Input
-                                    placeholder="Agustín Perez"
-                                    type="text"
-                                    id="name"
-                                    name="name"
-                                    value={data.name}
-                                    onChange={(e) =>
-                                        setData('name', e.target.value)
-                                    }
-                                    className="mt-1 block w-full"
-                                />
-                                <InputError message={errors.name} />
-                            </div>
-
-                            <div className="mt-3">
-                                <Label>Teléfono</Label>
-                                <InputHint
-                                    className="mt-2"
-                                    message="Un número de teléfono válido contiene sólo 10 dígitos"
-                                />
-                                <Input
-                                    placeholder="3804125834"
-                                    type="text"
-                                    id="phone"
-                                    name="phone"
-                                    value={data.phone}
-                                    onChange={(e) =>
-                                        setData('phone', e.target.value)
-                                    }
-                                    className="mt-1 block w-full"
-                                />
-                                <InputError
-                                    className="mt-2"
-                                    message={errors.phone}
-                                />
-                            </div>
-
-                            <div className="mt-3">
-                                <Label>Nombre del niño</Label>
-                                <Input
-                                    placeholder="Ej: Juan"
-                                    type="text"
-                                    id="child_name"
-                                    name="child_name"
-                                    value={data.child_name}
-                                    onChange={(e) =>
-                                        setData('child_name', e.target.value)
-                                    }
-                                    className="mt-1 block w-full"
-                                />
-                                <InputError message={errors.child_name} />
-                            </div>
-
-                            <div className="mt-3">
-                                <Label htmlFor="attended_photo_session">
-                                    ¿Asistió a la sesión de fotos?
-                                </Label>
-                                <div className="mt-2 flex gap-4">
-                                    <label className="flex cursor-pointer items-center">
-                                        <input
-                                            type="radio"
-                                            name="attended_photo_session"
-                                            value="true"
-                                            checked={
-                                                data.attended_photo_session ===
-                                                true
-                                            }
-                                            onChange={() =>
-                                                setData(
-                                                    'attended_photo_session',
-                                                    true,
-                                                )
-                                            }
-                                            className="mr-2"
-                                        />
-                                        Sí
-                                    </label>
-                                    <label className="flex cursor-pointer items-center">
-                                        <input
-                                            type="radio"
-                                            name="attended_photo_session"
-                                            value="false"
-                                            checked={
-                                                data.attended_photo_session ===
-                                                false
-                                            }
-                                            onChange={() =>
-                                                setData(
-                                                    'attended_photo_session',
-                                                    false,
-                                                )
-                                            }
-                                            className="mr-2"
-                                        />
-                                        No
-                                    </label>
-                                </div>
-                            </div>
-
-                            <div className="mt-6 flex flex-col justify-end gap-3 md:flex-row">
-                                <Button
-                                    variant="outline"
-                                    onClick={toStep('schools')}
-                                >
-                                    Anterior
-                                </Button>
-
-                                <Button onClick={toStep('products')}>
-                                    Siguiente
-                                </Button>
-                            </div>
-                        </AccordionContent>
-                    </AccordionItem>
-
-                    <AccordionItem value="products">
-                        <AccordionTrigger onClick={toStep('products')}>
-                            <div className="flex items-center gap-2">
-                                {error_count['products'] && (
-                                    <AlertCircle className="h-5 w-5 stroke-destructive" />
-                                )}
-                                Productos
-                                {data.order_details && (
-                                    <Badge className="ml-2">{`${data.order_details.length} productos`}</Badge>
-                                )}
-                            </div>
-                        </AccordionTrigger>
-                        <AccordionContent className="px-1">
-                            <Combobox
-                                items={combos.map((combo) => ({
-                                    label: combo.name,
-                                    value: combo.id,
-                                }))}
-                                action={(value) =>
-                                    handleAddCombo(Number(value))
-                                }
-                                open={comboDropdownOpen}
-                                setOpen={setComboDropdownOpen}
-                                placeholder="Buscar combo"
-                            >
-                                <Button variant="secondary" size="sm">
-                                    Añadir desde combo
-                                    <PlusIcon />
-                                </Button>
-                            </Combobox>
-
-                            <Combobox
-                                items={products.map((product) => ({
-                                    label: product.name,
-                                    value: product.id,
-                                }))}
-                                action={(value) =>
-                                    handleAddProduct(Number(value))
-                                }
-                                open={productDropdownOpen}
-                                setOpen={setProductDropdownOpen}
-                                placeholder="Buscar producto"
-                            >
-                                <Button
-                                    variant="secondary"
-                                    size="sm"
-                                    className="ml-2"
-                                >
-                                    Añadir producto
-                                    <PlusIcon />
-                                </Button>
-                            </Combobox>
-
-                            <ul className="my-2 gap-4">
-                                {data.order_details.map((selected, index) => {
-                                    const product = products.find(
-                                        (p) => p.id === selected.product_id,
-                                    );
-                                    const combo = combos.find(
-                                        (c) => c.id === selected.combo_id,
-                                    );
-
-                                    // The product may have been deleted since
-                                    // the detail was added or restored
-                                    if (!product) return null;
-
-                                    return (
-                                        <li
-                                            className="flex items-center justify-between rounded-md border border-input bg-background px-4 py-2"
-                                            key={`${product.id}-${combo ? combo.id : ''}-${index}`}
-                                        >
-                                            <div className="flex flex-col gap-2">
-                                                <span>
-                                                    {product.name}
-                                                    {combo
-                                                        ? ` (combo ${combo.name})`
-                                                        : ''}
-                                                </span>
-                                                {product.product_type_id ===
-                                                1 ? (
-                                                    <>
-                                                        <div className="flex items-center">
-                                                            <Badge
-                                                                variant="outline"
-                                                                className="gap-1 rounded-lg"
-                                                            >
-                                                                <Square
-                                                                    className="h-4 w-4"
-                                                                    style={{
-                                                                        fill: selected
-                                                                            .variant
-                                                                            ?.color,
-                                                                    }}
-                                                                />
-                                                            </Badge>
-                                                            <Badge
-                                                                variant="outline"
-                                                                className="rounded-lg"
-                                                            >
-                                                                {
-                                                                    selected
-                                                                        .variant
-                                                                        ?.background
-                                                                }
-                                                            </Badge>
-                                                            <Badge
-                                                                variant="outline"
-                                                                className="rounded-lg"
-                                                            >
-                                                                {selected
-                                                                    .variant
-                                                                    ?.photo_type ===
-                                                                'individual' ? (
-                                                                    <User className="h-4 w-4" />
-                                                                ) : (
-                                                                    <Users className="h-4 w-4" />
-                                                                )}
-                                                            </Badge>
-                                                            <Badge
-                                                                variant="outline"
-                                                                className="rounded-lg"
-                                                            >
-                                                                {selected
-                                                                    .variant
-                                                                    ?.orientation ===
-                                                                'vertical' ? (
-                                                                    <RectangleVertical className="h-4 w-4" />
-                                                                ) : (
-                                                                    <RectangleHorizontal className="h-4 w-4" />
-                                                                )}
-                                                            </Badge>
-                                                        </div>
-                                                    </>
-                                                ) : undefined}
-                                            </div>
-                                            <div className="flex gap-1">
-                                                <Button
-                                                    variant="warning"
-                                                    size="icon"
-                                                    title="Editar variantes y notas"
-                                                    onClick={(e) => {
-                                                        e.preventDefault();
-                                                        handleEditProduct(
-                                                            index,
-                                                        );
-                                                    }}
-                                                >
-                                                    <Edit className="h-4 w-4" />
-                                                </Button>
-                                                <Button
-                                                    variant="destructive"
-                                                    size="icon"
-                                                    title="Quitar del pedido"
-                                                    onClick={(e) => {
-                                                        e.preventDefault();
-                                                        handleRemoveProduct(
-                                                            index,
-                                                        );
-                                                    }}
-                                                >
-                                                    <Trash className="h-4 w-4" />
-                                                </Button>
-                                            </div>
-                                        </li>
-                                    );
-                                })}
-                            </ul>
-
-                            <InputError message={errors.order_details} />
-
-                            <div className="mt-6 flex flex-col justify-end gap-3 md:flex-row">
-                                <Button
-                                    variant="outline"
-                                    onClick={toStep('client')}
-                                >
-                                    Anterior
-                                </Button>
-
-                                <Button onClick={toStep('order')}>
-                                    Siguiente
-                                </Button>
-                            </div>
-                        </AccordionContent>
-                    </AccordionItem>
-
-                    <AccordionItem value="order">
-                        <AccordionTrigger onClick={toStep('order')}>
-                            <div className="flex items-center gap-2">
-                                {error_count['products'] && (
-                                    <AlertCircle className="h-5 w-5 stroke-destructive" />
-                                )}
-                                Pedido
-                            </div>
-                        </AccordionTrigger>
-                        <AccordionContent className="px-1">
-                            <div className="mt-3">
-                                <Label>Precio final</Label>
-                                <InputHint
-                                    message={`Calculado a base de (${Object.keys(counts.combos).length}) combo y (${counts.undefinedCount}) productos`}
-                                />
-                                <Input
-                                    type="number"
-                                    id="total_price"
-                                    name="total_price"
-                                    value={data.total_price}
-                                    onChange={(e) =>
-                                        setData('total_price', e.target.value)
-                                    }
-                                    className="mt-1 block w-full"
-                                />
-                                <InputError message={errors.total_price} />
-                            </div>
-
-                            <div className="mt-3">
-                                <Label>Cuotas</Label>
-                                <Input
-                                    type="number"
-                                    id="payment_plan"
-                                    name="payment_plan"
-                                    value={data.payment_plan}
-                                    onChange={(e) =>
-                                        setData('payment_plan', e.target.value)
-                                    }
-                                    className="mt-1 block w-full"
-                                />
-                                <InputError
-                                    className="mt-2"
-                                    message={errors.payment_plan}
-                                />
-                                <InputHint
-                                    className="mt-2"
-                                    message={`${data.payment_plan} cuotas de ${formatPrice(Number(data.total_price) / (Number(data.payment_plan) || 1))}`}
-                                />
-                            </div>
-
-                            <div className="mt-3">
-                                <Label>Primer vencimiento</Label>
-                                <div className="block">
-                                    <DatePicker
-                                        placeholder="Primer vencimiento"
-                                        date={
-                                            data.due_date
-                                                ? new Date(data.due_date)
-                                                : new Date()
-                                        }
-                                        setDate={(date) =>
-                                            setData(
-                                                'due_date',
-                                                format(
-                                                    date ?? new Date(),
-                                                    'yyyy-MM-dd',
-                                                ),
-                                            )
-                                        }
-                                    />
-                                </div>
-
-                                <InputError
-                                    className="mt-2"
-                                    message={errors.due_date}
-                                />
-                            </div>
-
-                            <div className="mt-6 flex flex-col justify-end gap-3 md:flex-row">
-                                <Button variant="outline" asChild>
-                                    <Link href={route('orders.index')}>
-                                        Cancelar
-                                    </Link>
-                                </Button>
-
-                                <Button
-                                    variant="outline"
-                                    disabled={processing}
-                                    onClick={handleSaveAsDraft}
-                                >
-                                    Guardar como borrador
-                                </Button>
-
-                                <Button
-                                    variant="secondary"
-                                    disabled={processing}
-                                    onClick={submit}
-                                >
-                                    Guardar
-                                </Button>
-
-                                <Button
-                                    disabled={processing}
-                                    onClick={handleSaveAndContinue}
-                                >
-                                    Guardar y seguir vendiendo
-                                </Button>
-                            </div>
-                        </AccordionContent>
-                    </AccordionItem>
+                    <SchoolStep
+                        form={form}
+                        schools={schools}
+                        schoolLevels={schoolLevels}
+                    />
+                    <ClientStep form={form} />
+                    <ProductsStep
+                        form={form}
+                        products={products}
+                        combos={combos}
+                    />
+                    <OrderStep form={form} />
                 </Accordion>
             </form>
         </AppLayout>
